requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-import-export==3.3.8
django-jazzmin==3.0.0
django-mathfilters==1.0.0
django-safedelete==1.3.3
django-taggit==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.7
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.11.0
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
